.history/travel-guide-site/requirements_20241210095128.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
pymysql==1.0.3
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0

